requirements_updated.txt
# Core Django Framework
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0

# Data Science & Analysis
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
scipy==1.14.1
scikit-learn==1.5.2
statsmodels==0.14.4

# Database
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
psycopg2-binary==2.9.10

# Excel/Data Processing
openpyxl==3.1.5
xlrd==2.0.1

# Visualization
plotly==5.24.1

# Big Data (Optional)
pyspark==3.5.3
dask==2024.11.2

# Geospatial (Optional)
geopandas==1.0.1
folium==0.18.0
shapely==2.0.6

# Image Processing & OCR
opencv-python==4.10.0.84
pytesseract==0.3.13
scikit-image==0.24.0
Pillow==11.0.0

# Security & Cryptography
cryptography==43.0.3

# NLP (Optional)
nltk==3.9.1
spacy==3.8.2
textblob==0.18.0.post0

# Web Frameworks (Optional)
streamlit==1.40.1
dash==2.18.2

# Machine Learning
joblib==1.4.2

# PDF Generation
reportlab==4.2.5

# File Storage
cloudinary==1.41.0

# Jupyter (Development)
jupyter==1.1.1
